+++ modules/Material/Device.qml
@@ -24,12 +24,14 @@ Object {
 
     //some kind of enum, by screen size
     property int type
+    
     readonly property int phone: 0
     readonly property int phablet: 1
     readonly property int tablet: 2
     readonly property int desktop: 3
     readonly property int tv: 4
     readonly property int unknown: 5 //it's either bigger than tv or smaller than phone
+    
     readonly property string name: {
         switch (type) {
             case 0:
@@ -52,6 +54,7 @@ Object {
                 break;
         }
     }
+    
     readonly property string iconName: {
         switch (type) {
             case 0:
@@ -74,4 +77,6 @@ Object {
                 break;
         }
     }
+
+    readonly property bool isMobile: type == phone || type == tablet
 }