+++ CHANGELOG.md
@@ -4,6 +4,7 @@ QML Material CHANGELOG
  * Add a ProgressCircle component following the Google spec
  * Add a tooltip component and use it in the ActionBar
  * [Tooltip] Clean up the code and improve the style
+ * [Tooltip] Clean up the code and implement desktop-sized tooltips
 
 ### Version 0.0.5 (March 4, 2015)